requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.9
flightsql-dbapi==0.2.1
influxdb==5.3.1
influxdb-client==1.36.1
pandas==1.5.0